refactor LegacySignedPublicKeyProto
